log in to edit workout

+++ frontend/src/pages/workoutDetails.jsx
@@ -1,5 +1,6 @@
 
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import {useAuthContext} from '../hooks/useAuthContext'
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
@@ -7,12 +8,20 @@ import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
 
-  const {dispatch} = useWorkoutsContext()
+  const { dispatch } = useWorkoutsContext()
+  const {user} = useAuthContext()
 
   
   const handleDelete = async () => {
+
+    if (!user) {
+      return
+    }
     const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'Authorization' : `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
